+++ src/plone/app/mosaic/upgrades.py
@@ -2,7 +2,10 @@
 from plone import api
 from plone.app.blocks.layoutbehavior import ILayoutAware
 from plone.app.mosaic.setuphandlers import create_ttw_layout_examples
+from plone.registry.interfaces import IRegistry
 from Products.CMFCore.utils import getToolByName
+from zope.component import getUtility
+from plone.app.mosaic.interfaces import IAction
 
 
 PROFILE_ID = 'profile-plone.app.mosaic:default'
@@ -138,3 +141,19 @@ def upgrade_to_1_0_0(context):
     for brain in catalog(object_provides=ILayoutAware.__identifier__):
         obj = brain.getObject()
         obj.reindexObject(idxs=['layout'])
+
+
+def upgrade_to_1_0rc3(context):
+    registry = getUtility(IRegistry)
+    actions = registry.get('plone.app.mosaic.default_available_actions', None)
+    if actions and 'remove' in actions:
+        actions.remove('remove')
+        registry.records['plone.app.mosaic.default_available_actions'] = actions
+
+    entry = registry.forInterface(
+        IAction, prefix='plone.app.mosaic.secondary_actions.remove', check=False)
+
+    for name in entry.__schema__.names():
+        key = entry.__prefix__ + name
+        if key in registry.records:
+            del registry.records[key]