+++ CHANGES.rst
@@ -7,6 +7,12 @@ Changelog
 - Remove testing for Plone 4.3. For Plone 4.3 support, please use
   plone.app.mosaic < 2.0.
 
+- Move tile remove button onto tile instead of in toolbar
+  [vangheem]
+
+- Fix modal to be structured like other plone modals
+  [vangheem]
+
 
 1.0 (2016-04-11)
 ----------------